+++ typescript/core_api_bindings/api_container_service_pb.d.ts
@@ -116,6 +116,48 @@ export namespace GenerateFilesResponse {
   }
 }
 
+export class LoadStaticFilesArgs extends jspb.Message {
+  getServiceId(): string;
+  setServiceId(value: string): void;
+
+  getStaticFilesMap(): jspb.Map<string, boolean>;
+  clearStaticFilesMap(): void;
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): LoadStaticFilesArgs.AsObject;
+  static toObject(includeInstance: boolean, msg: LoadStaticFilesArgs): LoadStaticFilesArgs.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: LoadStaticFilesArgs, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): LoadStaticFilesArgs;
+  static deserializeBinaryFromReader(message: LoadStaticFilesArgs, reader: jspb.BinaryReader): LoadStaticFilesArgs;
+}
+
+export namespace LoadStaticFilesArgs {
+  export type AsObject = {
+    serviceId: string,
+    staticFilesMap: Array<[string, boolean]>,
+  }
+}
+
+export class LoadStaticFilesResponse extends jspb.Message {
+  getCopiedStaticFileRelativeFilepathsMap(): jspb.Map<string, string>;
+  clearCopiedStaticFileRelativeFilepathsMap(): void;
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): LoadStaticFilesResponse.AsObject;
+  static toObject(includeInstance: boolean, msg: LoadStaticFilesResponse): LoadStaticFilesResponse.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: LoadStaticFilesResponse, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): LoadStaticFilesResponse;
+  static deserializeBinaryFromReader(message: LoadStaticFilesResponse, reader: jspb.BinaryReader): LoadStaticFilesResponse;
+}
+
+export namespace LoadStaticFilesResponse {
+  export type AsObject = {
+    copiedStaticFileRelativeFilepathsMap: Array<[string, string]>,
+  }
+}
+
 export class StartServiceArgs extends jspb.Message {
   getServiceId(): string;
   setServiceId(value: string): void;
@@ -205,6 +247,50 @@ export namespace PortBinding {
   export type AsObject = {
     interfaceIp: string,
     interfacePort: string,
+  }
+}
+
+export class GetServiceInfoArgs extends jspb.Message {
+  getServiceId(): string;
+  setServiceId(value: string): void;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): GetServiceInfoArgs.AsObject;
+  static toObject(includeInstance: boolean, msg: GetServiceInfoArgs): GetServiceInfoArgs.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: GetServiceInfoArgs, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): GetServiceInfoArgs;
+  static deserializeBinaryFromReader(message: GetServiceInfoArgs, reader: jspb.BinaryReader): GetServiceInfoArgs;
+}
+
+export namespace GetServiceInfoArgs {
+  export type AsObject = {
+    serviceId: string,
+  }
+}
+
+export class GetServiceInfoResponse extends jspb.Message {
+  getIpAddr(): string;
+  setIpAddr(value: string): void;
+
+  getSuiteExecutionVolumeMountDirpath(): string;
+  setSuiteExecutionVolumeMountDirpath(value: string): void;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): GetServiceInfoResponse.AsObject;
+  static toObject(includeInstance: boolean, msg: GetServiceInfoResponse): GetServiceInfoResponse.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: GetServiceInfoResponse, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): GetServiceInfoResponse;
+  static deserializeBinaryFromReader(message: GetServiceInfoResponse, reader: jspb.BinaryReader): GetServiceInfoResponse;
+}
+
+export namespace GetServiceInfoResponse {
+  export type AsObject = {
+    ipAddr: string,
+    suiteExecutionVolumeMountDirpath: string,
   }
 }
 
@@ -411,6 +497,26 @@ export namespace WaitForEndpointAvailabilityArgs {
     retries: number,
     retriesDelayMilliseconds: number,
     bodyText: string,
+  }
+}
+
+export class ExecuteBulkCommandsArgs extends jspb.Message {
+  getSerializedCommands(): string;
+  setSerializedCommands(value: string): void;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): ExecuteBulkCommandsArgs.AsObject;
+  static toObject(includeInstance: boolean, msg: ExecuteBulkCommandsArgs): ExecuteBulkCommandsArgs.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: ExecuteBulkCommandsArgs, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): ExecuteBulkCommandsArgs;
+  static deserializeBinaryFromReader(message: ExecuteBulkCommandsArgs, reader: jspb.BinaryReader): ExecuteBulkCommandsArgs;
+}
+
+export namespace ExecuteBulkCommandsArgs {
+  export type AsObject = {
+    serializedCommands: string,
   }
 }